// src/interfaces/nuevoFilm.response.ts
// Generated by https://quicktype.io

export interface NuevoFilmInterface {
    data:    Data;
}

export interface Data {
    nuevoFilm:  string;
    nombreFilm: string;
    persona:    string;
    planeta:    string;
    especie:    string;
    nave:       string;
    vehiculo:   string;
    emocion:    string;
    id:         number;
}